Move fixture to evaluation view tests module
